fhir: move to test-infra
