+++ fhir/src/main/resources/application.properties
@@ -15,9 +15,9 @@
 ## limitations under the License.
 ## ---------------------------------------------------------------------------
 
-serverUrl=http://localhost:8081/hapi-fhir-jpaserver/fhir
+serverUrl=http://localhost:8081/fhir
 
-fhirVersion=DSTU3
+fhirVersion=R4
 
 input=target/work/fhir/input
